requirements.txt
flask==1.0.2
requests==2.18.4
numpy==1.14.3
redis==3.2.1
Pillow==5.1.0
opencv-python==4.1.0.25
